backend/requirements.txt
# ZivonPay Backend Requirements
# Python 3.11+

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# QR Code Generation
qrcode[pil]==7.4.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.39.2

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
pytz==2024.1
ujson==5.9.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
